add update component form
