+++ components/Form.js
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+import { useState } from "react";
 
 const FormContainer = styled.div`
   display: flex;
@@ -42,13 +43,48 @@ const FormButton = styled.button`
     color: white;
   }
 `;
-
 export default function Form() {
+  const [comment, setComment] = useState("");
+
+  const endpoint = "/api/form"; //API endpoint where we send form data
+  /*
+  const [comments, setComments] = useState([]);
+  const fetchComments = async () => {
+    const response = await fetch(endpoint);
+    const result = await response.json();
+    setComments(result);
+  };
+*/
+  const handleSubmit = async (event) => {
+    event.preventDefault();
+    const dataForm = {
+      recipe: event.target.recipe.value,
+      name: event.target.name.value,
+      comment: event.target.comment.value,
+    };
+    const JSONdata = JSON.stringify(dataForm); //send data to server in JSON format
+
+    const options = {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSONdata, //JSON data we send above
+    };
+    const response = await fetch(endpoint, options); //send the form data to API and Vercel and get a response
+    const result = await response.json(); //convert response it into json
+    setComment(result);
+    console.log(result);
+  };
+  const comments = JSON.parse(localStorage.getItem("comments")) || [];
+  comments.push(dataForm);
+  localStorage.setItem("comments", JSON.stringify(comments));
+
   return (
     <>
       <FormContainer>
         <FormTitle>You may leave a comment here:</FormTitle>
-        <form action="/api/form" method="post">
+        <form action="/api/form" method="post" onSubmit={handleSubmit}>
           <FormField>
             <label htmlFor="recipe">Recipe:</label>
             <input
