+++ package.json
@@ -19,7 +19,8 @@
     "react-icons": "^4.8.0",
     "styled-components": "^5.3.8",
     "swr": "^2.1.0",
-    "uid": "^2.0.1"
+    "uid": "^2.0.1",
+    "use-local-storage-state": "^18.2.1"
   },
   "devDependencies": {
     "@svgr/webpack": "^6.5.1",
